+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Version 0.2.0
+
+ - Internal Changes on the test runner
+ - Events can now be attached via $element.on, using jquery custom events.
+ - Update comment banner
+ - Update Grunt to 0.4.x
+ - Fix dependencies
+
 ## Version 0.1.4
 
  - Fix comments.